Split block_certificates into a new module file.
