+++ src/v2/mod.rs
@@ -6,9 +6,7 @@ use crate::{
     endpoints, id,
     id::types::AccountCredentialMessage,
     types::{
-        self,
-        block_certificates::*,
-        hashes,
+        self, block_certificates, hashes,
         hashes::{BlockHash, TransactionHash, TransactionSignHash},
         smart_contracts::{
             ContractContext, InstanceInfo, InvokeContractResult, ModuleReference, WasmModule,
@@ -2195,13 +2193,13 @@ impl Client {
     pub async fn get_block_certificates(
         &mut self,
         bi: impl IntoBlockIdentifier,
-    ) -> endpoints::QueryResult<QueryResponse<BlockCertificates>> {
+    ) -> endpoints::QueryResult<QueryResponse<block_certificates::BlockCertificates>> {
         let response = self
             .client
             .get_block_certificates(&bi.into_block_identifier())
             .await?;
         let block_hash = extract_metadata(&response)?;
-        let response = BlockCertificates::try_from(response.into_inner())?;
+        let response = block_certificates::BlockCertificates::try_from(response.into_inner())?;
         Ok(QueryResponse {
             block_hash,
             response,
